Initialise several class members with a memset to null
All of them were found with Coverity Scan

Generally they're unlikely to cause problems as their current use in the
API and RecastDemo will initialise the members before first use, but
explicitly initialising them is clearly safer. Eventually they should just
use normal constructors rather than memset, but there is a lot of code that
would need to change, and in a breaking way, so this is a good interim fix.